Add TaskViewFactory
