+++ ground/src/main/java/com/google/android/ground/ui/datacollection/components/TaskView.kt
@@ -15,12 +15,8 @@
  */
 package com.google.android.ground.ui.datacollection.components
 
-import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.annotation.DrawableRes
-import androidx.annotation.StringRes
-import androidx.appcompat.content.res.AppCompatResources
 import androidx.fragment.app.Fragment
 import com.google.android.ground.databinding.TaskFragWithHeaderBinding
 import com.google.android.ground.databinding.TaskFragWithoutHeaderBinding
@@ -57,13 +53,6 @@ data class TaskViewWithHeader(private val binding: TaskFragWithHeaderBinding) : 
   override fun addTaskView(view: View) {
     binding.taskContainer.addView(view)
   }
-
-  companion object {
-    fun create(layoutInflater: LayoutInflater): TaskView {
-      val binding = TaskFragWithHeaderBinding.inflate(layoutInflater)
-      return TaskViewWithHeader(binding)
-    }
-  }
 }
 
 /** Implementation of [TaskView] without an embedded header. */
@@ -80,21 +69,5 @@ data class TaskViewWithoutHeader(private val binding: TaskFragWithoutHeaderBindi
 
   override fun addTaskView(view: View) {
     binding.taskContainer.addView(view)
-  }
-
-  companion object {
-    fun create(
-      layoutInflater: LayoutInflater,
-      @DrawableRes iconResId: Int? = null,
-      @StringRes labelResId: Int? = null
-    ): TaskView {
-      val binding = TaskFragWithoutHeaderBinding.inflate(layoutInflater)
-      iconResId?.let {
-        val drawable = AppCompatResources.getDrawable(layoutInflater.context, it)
-        binding.headerIcon.setImageDrawable(drawable)
-      }
-      labelResId?.let { binding.headerLabel.setText(labelResId) }
-      return TaskViewWithoutHeader(binding)
-    }
   }
 }
